requirements.txt
mcp>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.0
